# [nnc] Use minimally-sized rfactor buffer to improve performance