Stack from [ghstack](https://github.com/ezyang/ghstack):
* **#60161 [nnc] Use minimally-sized rfactor buffer to improve performance**
* #60160 [nnc][tests] Tests and benchmarks for computeSum

When rfactoring a reduction nested inside an output axis, the rfactor
buffer was created with size equal to the map axis, which is too big.  We only
want to fill the rfactor buffer along the rfactor axis, and then fill the
output axis with the final reduction.

Differential Revision: [D29190421](https://our.internmc.facebook.com/intern/diff/D29190421/)