@bertmaher 2021-06-17
Perf results didn't make it from phabricator so here they are:
Before:
```
------------------------------------------------------------------------------------------
Benchmark                                   Time           CPU Iterations UserCounters...
------------------------------------------------------------------------------------------
Reduce2DRow/OpSchedule/8/2097152/1    4741485 ns    4741553 ns          3 BYTES=14.1534G/s
Reduce2DRow/OpSchedule/64/262144/1    4824404 ns    4824483 ns          3 BYTES=13.9101G/s
Reduce2DRow/OpSchedule/4096/4096/1    7329771 ns    7329452 ns          2 BYTES=9.15829G/s
Reduce2DRow/OpSchedule/262144/64/1 1040878288 ns 1040849770 ns          1 BYTES=65.4825M/s
```
After:
```
------------------------------------------------------------------------------------------
Benchmark                                   Time           CPU Iterations UserCounters...
------------------------------------------------------------------------------------------
Reduce2DRow/OpSchedule/8/2097152/1    5304754 ns    5304568 ns          3 BYTES=12.6512G/s
Reduce2DRow/OpSchedule/64/262144/1    5342981 ns    5342843 ns          3 BYTES=12.5606G/s
Reduce2DRow/OpSchedule/4096/4096/1    4944474 ns    4944442 ns          3 BYTES=13.5759G/s
Reduce2DRow/OpSchedule/262144/64/1    5741781 ns    5743658 ns          2 BYTES=11.8666G/s
```