@bertmaher 2021-06-17
This change is not needed.  Calling compressBuffer after rfactor has the same effect.